refactor(issuance): rename client.rs to client_orig.rs for backup
