+++ cloud-wallet-openid4vc/src/issuance/mod.rs
@@ -6,6 +6,7 @@ pub mod authz_response;
 pub mod authz_server_metadata;
 pub mod claim_path_pointer;
 pub mod client;
+pub mod client_orig;
 pub mod credential_configuration;
 pub mod credential_formats;
 pub mod credential_offer;
